requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
black==23.12.1
boto3==1.34.11
botocore==1.34.11
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0
django-appconf==1.0.6
django-imagekit==5.0.0
django-js-asset==2.2.0
django-mptt==0.15.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-spectacular==0.27.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pilkit==3.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.16.2
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
